composer added
This was because I forgot to add "composer" to phansible's build out. However, having looked at it, I prefer the flexibility fo the composer role in https://github.com/geerlingguy/ansible-role-composer

The issue with that is it's full of ancillary b.s. that keeps it from installing unless you do his entire setup, so instead of using a git module, I just copied the important parts. source.txt has an explanation.

- also removed xhprof and inclued from the cli interface

Signed-off-by: terry chay <tychay@php.net>